chore: Remove EditPolicy-related GraphQL implementation
And also clean up some tests like hooks.test.js.
