+++ src/Mutations/index.js
@@ -1,2 +1,1 @@
 export * from './graphql/mutations';
-export { default as usePolicy } from './hooks/usePolicy';
